fix: Homeコンポーネントの関数の責務分離
