+++ virtual_partner_front/src/features/Home/components/HomeContainer.tsx
@@ -1,0 +1,16 @@
+import { HomeView } from "./HomeView";
+import { useHome } from "../hooks/useHome";
+
+export const HomeContainer = () => {
+  const { modelPath, charaId, llmResponse, comments, handleResult } = useHome();
+
+  return (
+    <HomeView
+      modelPath={modelPath}
+      charaId={charaId}
+      llmResponse={llmResponse}
+      comments={comments}
+      onResult={handleResult}
+    />
+  );
+};
