+++ virtual_partner_front/src/features/Home/index.ts
@@ -1,0 +1,2 @@
+import { HomeContainer } from "./components/HomeContainer";
+export { HomeContainer as Home };
